backend/requirements.txt
Baker
black
django
django-extensions
django-filter
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-extra-fields
drf-spectacular
drf-writable-nested
gunicorn
model-bakery
psycopg2-binary
pylint
pylint-django
pytest
pytest-django
requests
